backend/requirements.txt
asgiref
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.10
dj-database-url==0.5.0
Django
django-allauth
django-debug-toolbar==3.2.4
django-filter==21.1
django-jazzmin==2.4.8
django-silk==4.2.0
djangorestframework
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
gprof2dot==2021.2.21
gunicorn==20.1.0
uvicorn[standard]
idna==3.3
importlib-metadata==4.10.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
openpyxl==3.0.9
psycopg2-binary==2.9.2
psycopg2
pycodestyle==2.7.0
Pygments==2.11.1
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==1.0.1
pytz==2021.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
urllib3==1.26.7
whitenoise==5.3.0
XlsxWriter==3.0.2
zipfile36==0.1.3
zipp==3.6.0
django-rest-passwordreset
Pillow
channels
django-debug-toolbar
channels-redis